Move mondo object models into lib.models.mondo
