+++ test/reference_data/test_mondo.py
@@ -4,6 +4,7 @@ from typing import Any
 
 import pytest
 
+from lib.models import MondoDiseaseContext, MondoDiseaseScope, MondoLinkingTarget
 from lib.reference_data import mondo
 
 
@@ -179,12 +180,12 @@ def test_get_mondo_parents_and_children_return_related_terms(
 def test_mondo_agent_message_includes_scoped_target() -> None:
     from lib.agents.mondo_linking_agent import build_mondo_agent_message
 
-    target = mondo.MondoLinkingTarget(
-        scope=mondo.MondoDiseaseScope.OCCURRENCE,
+    target = MondoLinkingTarget(
+        scope=MondoDiseaseScope.OCCURRENCE,
         paper_id=123,
         patient_variant_occurrence_id=456,
         disease_text='Marfan syndrome (MONDO:0007947)',
-        context=mondo.MondoDiseaseContext(
+        context=MondoDiseaseContext(
             gene_symbol='FBN1',
             occurrence_disease_text='Marfan syndrome (MONDO:0007947)',
         ),
